sub-metre-elevation
Need my AronaLayers-gen mod & the following JVM argument then start a new world :
-Dtellus.subMetreElevation=true

+++ src/main/java/com/yucareux/tellus/world/data/elevation/ShortRaster.java
@@ -2,22 +2,31 @@ package com.yucareux.tellus.world.data.elevation;
 
 import java.util.Arrays;
 
+/**
+ * Elevation raster.
+ *
+ * <p>Historically backed by {@code short} (whole-metre) samples; now backed by {@code float}
+ * so the sub-metre precision decoded from Terrarium tiles can be preserved when the
+ * sub-metre-elevation option is enabled. The class name is retained to avoid a wide rename.
+ * {@link #get} returns a {@code double} and {@link #set} accepts a {@code double}; when the
+ * option is off, callers store whole-metre (rounded) values, so behaviour is unchanged.
+ */
 final class ShortRaster {
    private final int width;
    private final int height;
-   private final short[] data;
+   private final float[] data;
 
-   private ShortRaster(int width, int height, short[] data) {
+   private ShortRaster(int width, int height, float[] data) {
       this.width = width;
       this.height = height;
       this.data = data;
    }
 
    static ShortRaster create(int width, int height) {
-      return new ShortRaster(width, height, new short[checkedSampleCount(width, height)]);
+      return new ShortRaster(width, height, new float[checkedSampleCount(width, height)]);
    }
 
-   static ShortRaster wrap(int width, int height, short[] data) {
+   static ShortRaster wrap(int width, int height, float[] data) {
       if (data.length != checkedSampleCount(width, height)) {
          throw new IllegalArgumentException("Invalid raster buffer");
       } else {
@@ -44,15 +53,15 @@ final class ShortRaster {
       return this.height;
    }
 
-   short get(int x, int y) {
+   double get(int x, int y) {
       return this.data[x + y * this.width];
    }
 
-   void set(int x, int y, short value) {
-      this.data[x + y * this.width] = value;
+   void set(int x, int y, double value) {
+      this.data[x + y * this.width] = (float)value;
    }
 
-   void fill(short value) {
-      Arrays.fill(this.data, value);
+   void fill(double value) {
+      Arrays.fill(this.data, (float)value);
    }
 }
